# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "The Tower API (v1)",
        "description": "Endpoints for various resources in The Tower.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Version 1"
        }
    ],
    "paths": {
...
        "/profileBanners/{id}": {
            "get": {
                "tags": [
                    "ProfileBanners"
                ],
                "summary": "Get a profileBanner",
                "description": "Retrieve a single profileBanner.",
                "operationId": "2c284d72438c8a699b0fa3d5b21d8ab8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ProfileBanner detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProfileBannerResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ProfileBanners"
                ],
                "summary": "Delete profileBanner",
                "description": "Soft-delete a profileBanner.",
                "operationId": "c331ed96990d64776bb17e6ebd4a9b59",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "ProfileBanners"
                ],
                "summary": "Update profileBanner",
                "description": "Update an existing profileBanner.",
                "operationId": "6620de2108f88e54bbfb1a89331e2e89",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProfileBannerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated profileBanner",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProfileBannerResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/profileBanners": {
            "get": {
                "tags": [
                    "ProfileBanners"
                ],
                "summary": "List profileBanners",
                "description": "List profileBanners (paginated).",
                "operationId": "e51086f79727f055aa35e51bccedaec9",
                "responses": {
                    "200": {
                        "description": "Paginated list of profileBanners",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProfileBannerCollection"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "ProfileBanners"
                ],
                "summary": "Create profileBanner",
                "description": "Create a new profileBanner.",
                "operationId": "1a2695d49d7ef308d5ddd1f2e954b31b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProfileBannerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProfileBannerResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/profileBanners/{id}/restore": {
            "patch": {
                "tags": [
                    "ProfileBanners"
                ],
                "summary": "Restore profileBanner",
                "description": "Restore a soft-deleted profileBanner.",
                "operationId": "7b6eb65b9a3156891413c9ba9acd33a9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restored profileBanner",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProfileBannerResource"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "ProfileBanner is not deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreProfileBannerRequest": {
                "description": "Payload sent when creating a new ProfileBanner",
                "required": [
                    "name",
                    "value"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 191,
                        "example": "Epic Battle Theme"
                    },
                    "value": {
                        "type": "number",
                        "format": "float",
                        "example": 0.006
                    }
                },
                "type": "object"
            },
...
            "UpdateProfileBannerRequest": {
                "description": "Payload for updating an existing ProfileBanner. All fields optional.",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 191,
                        "example": "Epic Battle Theme Remix"
                    },
                    "value": {
                        "type": "number",
                        "format": "float",
                        "example": 0.007
                    }
                },
                "type": "object"
            },
...
            "ProfileBannerCollection": {
                "description": "Paginated list of ProfileBanners",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProfileBannerResource"
                        }
                    },
                    "links": {
                        "properties": {
                            "first": {
                                "type": "string",
                                "format": "uri",
                                "example": "/api/v1/profile-banners?page=1"
                            },
                            "last": {
                                "type": "string",
                                "format": "uri",
                                "example": "/api/v1/profile-banners?page=2"
                            },
                            "prev": {
                                "type": "string",
                                "format": "uri",
                                "example": null,
                                "nullable": true
                            },
                            "next": {
                                "type": "string",
                                "format": "uri",
                                "example": "/api/v1/profile-banners?page=2",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "meta": {
                        "properties": {
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "from": {
                                "type": "integer",
                                "example": 1
                            },
                            "last_page": {
                                "type": "integer",
                                "example": 2
                            },
                            "path": {
                                "type": "string",
                                "format": "uri",
                                "example": "/api/v1/profile-banners"
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 50
                            },
                            "to": {
                                "type": "integer",
                                "example": 50
                            },
                            "total": {
                                "type": "integer",
                                "example": 100
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ProfileBannerResource": {
                "description": "Represents a ProfileBanner available in the API",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Mech World"
                    },
                    "value": {
                        "type": "number",
                        "format": "float",
                        "example": 0.006
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}